Updated PR: code refactoring, enhance tests
