+++ src/braket/ahs/driving_field.py
@@ -14,6 +14,7 @@
 from __future__ import annotations
 
 from typing import List, Union
+import math
 
 from braket.ahs.discretization_types import DiscretizationProperties
 from braket.ahs.field import Field
@@ -98,9 +99,8 @@ class DrivingField(Hamiltonian):
             detunings (List[float]): The values of the detuning
             phases (List[float]): The values of the phase
         """
-        assert len(times) == len(amplitudes)
-        assert len(times) == len(detunings)
-        assert len(times) == len(phases)
+        if not (len(times) == len(amplitudes) == len(detunings) == len(phases)):
+            raise ValueError("The length of the list for time points, amplitudes, detunings and phases are not equal.")
 
         amplitude = TimeSeries()
         detuning = TimeSeries()
@@ -171,3 +171,36 @@ class DrivingField(Hamiltonian):
         return DrivingField(
             amplitude=discretized_amplitude, phase=discretized_phase, detuning=discretized_detuning
         )
+
+    @staticmethod
+    def rabi_pulse(
+        rabi_pulse_area: float, omega_max: float, omega_slew_rate_max: float
+    ) -> TimeSeries:
+        """Get a time series for Rabi frequency with specified Rabi phase, maximum amplitude
+        and maximum slew rate
+            Args:
+                rabi_pulse_area (float): Total area under the Rabi frequency time series
+                omega_max (float): The maximum amplitude
+                omega_slew_rate_max (float): The maximum slew rate
+            Returns:
+                TimeSeries: time series for the time dependent Rabi frequency
+            Notes: Rabi phase is equal to the integral of the amplitude of a time-dependent
+                Rabi frequency Omega(t).
+        """
+
+        phase_threshold = omega_max**2 / omega_slew_rate_max
+        if rabi_pulse_area <= phase_threshold:
+            t_ramp = math.sqrt(rabi_pulse_area / omega_slew_rate_max)
+            t_plateau = 0
+        else:
+            t_ramp = omega_max / omega_slew_rate_max
+            t_plateau = (rabi_pulse_area / omega_max) - t_ramp
+        t_pulse = 2 * t_ramp + t_plateau
+        time_points = [0, t_ramp, t_ramp + t_plateau, t_pulse]
+        amplitude_values = [0, t_ramp * omega_slew_rate_max, t_ramp * omega_slew_rate_max, 0]
+
+        ts = TimeSeries()
+        for t, v in zip(time_points, amplitude_values):
+            ts.put(t, v)
+
+        return ts
